refactor(checkout): CHECKOUT-7107 Address PR comments
