+++ packages/core/src/payment-integration/default-payment-integration-service.spec.ts
@@ -13,7 +13,7 @@ import {
     CheckoutActionCreator,
 } from "../checkout";
 import { DataStoreProjection } from "../common/data-store";
-import HostedForm from "../hosted-form/hosted-form";
+import { HostedForm, HostedFormFactory } from "../hosted-form";
 import { OrderActionCreator } from "../order";
 import { getOrder } from "../order/orders.mock";
 import PaymentActionCreator from "../payment/payment-action-creator";
@@ -27,6 +27,7 @@ import PaymentIntegrationStoreProjectionFactory from "./payment-integration-stor
 
 describe("DefaultPaymentIntegrationService", () => {
     let subject: PaymentIntegrationService;
+    let hostedFormFactory: HostedFormFactory;
     let paymentIntegrationSelectors: PaymentIntegrationSelectors;
     let internalCheckoutSelectors: InternalCheckoutSelectors;
     let store: Pick<CheckoutStore, "dispatch" | "getState">;
@@ -61,6 +62,7 @@ describe("DefaultPaymentIntegrationService", () => {
     let paymentActionCreator: Pick<PaymentActionCreator, "submitPayment">;
 
     beforeEach(() => {
+        hostedFormFactory = new HostedFormFactory(store as CheckoutStore);
         paymentIntegrationSelectors = {} as PaymentIntegrationSelectors;
 
         internalCheckoutSelectors = {
@@ -132,6 +134,7 @@ describe("DefaultPaymentIntegrationService", () => {
             store as CheckoutStore,
             storeProjectionFactory as PaymentIntegrationStoreProjectionFactory,
             checkoutActionCreator as CheckoutActionCreator,
+            hostedFormFactory as HostedFormFactory,
             orderActionCreator as OrderActionCreator,
             billingAddressActionCreator as BillingAddressActionCreator,
             consignmentActionCreator as ConsignmentActionCreator,
